Moved implementation of ray_tracing_2D to its own module
