+++ NuRadioMC/SignalProp/AnalyticRayTracing/install.sh
